#456921 [C#テキスト]4章演習問題
Practice4-1-3のコメントを修正しました。
プロパティで21世紀かのプロパティを読み取り専用に変更しました。
