+++ 章末問題4/Practice4-1-3/Program4-1-3.cs
@@ -1,6 +1,6 @@
 ﻿namespace Practice4_1_3 {
     class Program {
-        //YearMonthクラスに、1ヶ月後を求めるAddOneメソッドを追加してください。この時、自分自身のプロパティは変更せずに、新たなYearMonthオブジェクトを生成しその値を返してください。12月の時の処理に注意してください。
+        //YearMonthクラスに、1ヶ月後を求めるAddOneMonthメソッドを追加してください。この時、自分自身のプロパティは変更せずに、新たなYearMonthオブジェクトを生成しその値を返してください。12月の時の処理に注意してください。
         static void Main(string[] args) {
         }
     }
