+++ 章末問題4/Practice4-1-4/YearMonth.cs
@@ -18,7 +18,6 @@
             get {
                 return Year >= 2001 && Year <= 2100;
             }
-            set { Is21stCentury = value; }
         }
         /// <summary>
         /// YearMonthのコンストラクタ
@@ -39,9 +38,8 @@
 
             if (wNewMonth > 12) {
                 wNewMonth = 1;
-                wNewYear++;
+                ++wNewYear;
             }
-
             return new YearMonth(wNewYear, wNewMonth);
         }
         /// <summary>
